==> notebooks/00_create_graphs.ipynb <==
# Graph creation

%load_ext autoreload
%autoreload 2
import requests
import os
from dotenv import load_dotenv
from datetime import datetime
import time
import pandas as pd
from datasets import load_dataset
import sys
import json
load_dotenv()


from utils import get_sql_query

GRAPHS_URL = "https://graphs-api.codegpt.co" # http://localhost:8002
SECRET_KEY = os.environ['SECRET_KEY']

data_template = {
    "git_provider": "github",
    "repo_org": "myOrg",
    "repo_name": "myRepo",
    "branch": "myBranch",
    "connection_id": os.environ['CONNECTION_ID'],
    "generate_documentation": True,
}

# TEST RESPONSE
response = requests.get(GRAPHS_URL)
response.json()

# We use the Lite version of the SWE-bench dataset
swe_ds = load_dataset("princeton-nlp/SWE-bench_Lite", cache_dir="../data")
swe_df = pd.DataFrame(swe_ds['test'])
print(swe_df.shape)
swe_df.head()

# Get current graphs for the swe-bench user 
sql = f"""SELECT g.id, g.repo_id, g.status, r.commit_hash, r.repo_org, r.repo_name, r.created_at, r.size
FROM graphs g
JOIN repositories r
ON r.id = g.repo_id
WHERE user_id = '{os.environ['USER_ID']}'
--AND r.language = 'py'
"""

db_info = (get_sql_query(sql, ['graph_id', 'repo_id', 'status', 'base_commit', 'repo_org', 'repo_name', 'created_at', 'size'])
           .assign(repo = lambda x: x['repo_org'] + '/' + x['repo_name'])
           [['graph_id', 'repo_id', 'base_commit', 'repo', 'status', 'created_at', 'size']]
           )
# now = datetime.now().strftime('%Y%m%d_%H%M%S')
db_info.to_csv(f'../data/db_user_info.csv', index=False)
print(db_info.shape)
db_info

# graphs by status
db_info.value_counts('status')

# graphs by status
db_info.value_counts(['repo', 'status']).head(12)

merged_df = swe_df[['repo', 'instance_id', 'base_commit']].merge(db_info, on=['repo', 'base_commit'], how='left')
# From the merged dataframe we extract those without graph_id
repos = merged_df[(merged_df['graph_id'].isnull())]
print("Repos without graph ID:", repos.shape[0])

# Since the graph creation process uses previous commits as reference to accelerate the process,
# we will send a unique request for each repository.
# Once these graphs are created, we can run this notebook again
repos_unique = (pd.concat([
    # repos.drop_duplicates(subset=['repo'], keep='first'),
    repos.reset_index(drop=True).sample(frac=1),
    # repos.drop_duplicates(subset=['repo'], keep='last'),
    ]).drop_duplicates(subset=['instance_id'])
)
repos_unique

# A possible error here is that the secret_key is not valid or expired.
max_attempts = 1

for i, r in repos_unique.iterrows():
    attempts = 0
    data = data_template.copy()
    while attempts < max_attempts:
        if attempts == max_attempts:
            print(f"Failed to create graph {i+1} out of {len(repos_unique)}")
            break
        attempts += 1
        repo_org, repo_name = r['repo'].split('/')

        data['repo_org'] = repo_org
        data['repo_name'] = repo_name
        data['branch'] = r['base_commit']
        
        response = requests.post(f"{GRAPHS_URL}/v1/repo", 
                         headers={"Authorization": f"Bearer {SECRET_KEY}"},
                         json=data)
        if response.ok:
            print(r['repo'], response.json())
            break
        else:
            if response.status_code == 504: # Timeout, the graph is still being created
                break
            print(response)
    # wait 120 seconds
    time.sleep(120)
