==> notebooks/01_test_prompt.ipynb <==
## Test Patch 

**IMPORTANT:** You will need to start the API locally to run this notebook
- [Clone the repository](https://github.com/JudiniLabs/api-vicuna-deno)
- `cd api-vicuna-deno`
- [Install Deno](https://docs.deno.com/runtime/getting_started/installation/)
- Get the envs
- `deno run dev`


import requests
import os
from dotenv import load_dotenv
import pandas as pd
from datasets import load_dataset
load_dotenv()

API_URL = "http://localhost:8000" # https://api.codegpt.co
SECRET_KEY = os.environ['SECRET_KEY']

data_template = {
    "agentId": os.environ['AGENT_ID'], # This is an agent without prompt
    "graphId": "myGraphId",
    "stream": False,
    "format": "text",
    "messages": [],
    "model": 'llama-4-maverick-17b-groq' # claude-4-opus-vertex
}

# TEST RESPONSE
response = requests.get(API_URL)
response.text

# We use the Lite version of the SWE-bench dataset
swe_ds = load_dataset("princeton-nlp/SWE-bench_Lite", cache_dir="../data")
swe_df = pd.DataFrame(swe_ds['test'])
print(swe_df.shape)
swe_df.head()

db_info = pd.read_csv('../data/db_user_info.csv', parse_dates=['created_at']).query('status == "completed"')
db_info.head()

# INNER JOIN
merged_df = swe_df.merge(db_info, on=['repo', 'base_commit'], how='inner')[['repo', 'base_commit', 'problem_statement', 'hints_text', 'graph_id']]
merged_df.head()

row = merged_df.iloc[42]
graph_id = row['graph_id']
print(f"Repo: {row['repo']}")
print("=====================")
prompt = f"<issue>\n{row['problem_statement']}\n</issue>"
prompt += f"\n\n<hints>\n{row['hints_text']}\n</hints>" if row['hints_text'] else ""
print(prompt)

data = data_template
data['graphId'] = graph_id
data['format'] = 'json'
data['messages'] = [
    {
        "role": "user",
        "content": prompt
    }
]

response = requests.post(f"{API_URL}/api/v1/chat/completions", 
                    headers={"Authorization": f"Bearer {SECRET_KEY}"},
                    json=data)
if response.ok:
    messages = response.json()['choices']

else:
    print(response.json())
    
final_answer = messages[-1]['content']


print(final_answer)
