==> notebooks/02_get_patches.ipynb <==
## Test Patch 

**IMPORTANT:** You will need to start the API locally to run this notebook
- [Clone the repository](https://github.com/JudiniLabs/api-vicuna-deno)
- `cd api-vicuna-deno`
- [Install Deno](https://docs.deno.com/runtime/getting_started/installation/)
- Get the envs
- `deno run dev`


%load_ext autoreload
%autoreload 2

import requests
import os
from dotenv import load_dotenv
import pandas as pd
from datasets import load_dataset

load_dotenv()

API_URL = "http://localhost:8000" # https://api.codegpt.co
SECRET_KEY = os.environ['SECRET_KEY']

data_template = {
    "agentId": os.environ['AGENT_ID'], # This is an agent without prompt
    "graphId": "myGraphId",
    "stream": False,
    "format": "text",
    "messages": [],
}

# TEST RESPONSE
response = requests.get(API_URL)
response.text

# We use the Lite version of the SWE-bench dataset
swe_ds = load_dataset("princeton-nlp/SWE-bench_Lite", cache_dir="../data")
swe_df = pd.DataFrame(swe_ds['test'])
print(swe_df.shape)
swe_df.head(10)

db_info = pd.read_csv('../data/db_user_info.csv', parse_dates=['created_at']).query('status == "completed"')
db_info.head()

# INNER JOIN
merged_df = swe_df.merge(db_info, on=['repo', 'base_commit'], how='inner').set_index('instance_id')
print(merged_df.shape)
merged_df.head()

from CodeGPT_Patcher import CodeGPT_Patcher
import concurrent.futures

# import json
# import random
# SWE_PATH = os.environ['SWE_PATH']
# with open(f'{SWE_PATH}/CodeGPT_MultiAgent_Sonnet3_5.validate.json', 'r') as f:
#     prev_results = json.load(f)
# random_ids = random.sample(prev_results['unresolved_ids'], 30)
# random_ids[0]

# instance_ids = merged_df.sample(30).index
# instance_ids

import json

# I've updated db_info, so now we must use instance_id
instance_ids = ['django__django-15781', 'django__django-12125',
       'scikit-learn__scikit-learn-11040', 'sympy__sympy-21614',
       'sphinx-doc__sphinx-8595', 'django__django-12856',
       'django__django-17087', 'matplotlib__matplotlib-26011',
       'django__django-12589', 'pydata__xarray-4094',
       'pylint-dev__pylint-6506', 'scikit-learn__scikit-learn-25747',
       'django__django-14382', 'psf__requests-3362', 'sympy__sympy-23262',
       'sympy__sympy-13895', 'django__django-11049', 'pydata__xarray-4493',
       'matplotlib__matplotlib-23987', 'sympy__sympy-16106',
       'django__django-15498', 'pytest-dev__pytest-5227', 'sympy__sympy-21847',
       'sympy__sympy-13773', 'django__django-14997', 'sympy__sympy-19007',
       'pytest-dev__pytest-5221', 'sphinx-doc__sphinx-8721',
       'sympy__sympy-13031', 'scikit-learn__scikit-learn-13497']

# RUN_NAME = 'CodeGPT_MultiAgent_Vertex3_7'
# RUN_NAME = 'CodeGPT_MultiAgent_Opus4'
RUN_NAME = 'CodeGPT_MultiAgent_Llama4_Maverick'

run_path = f"../data/results/{RUN_NAME}"
os.makedirs(run_path, exist_ok=True)


# instance_ids = merged_df.sample(20).index
# instance_ids = [merged_df.iloc[0].name]
patchers = []

for instance_id in instance_ids:
    patchers.append(CodeGPT_Patcher(merged_df.loc[instance_id], API_URL, SECRET_KEY,  'llama-4-maverick-17b-groq')) # claude-3.7-vertex, claude-4-opus-vertex


def do_work(patcher):

    os.makedirs(f"{run_path}/errors", exist_ok=True)
    try:    
        has_diff = os.path.exists(f'{run_path}/diffs/{patcher.instance_id}.diff')
        has_trajs = os.path.exists(f'{run_path}/trajs/{patcher.instance_id}.json')
        has_error = os.path.exists(f'{run_path}/errors/{patcher.instance_id}.error')
        has_patches = os.path.exists(f'{run_path}/patches/{patcher.instance_id}_patches.json')
        if has_diff:
            print(f'{patcher.instance_id} has already been processed. Skipping it...')
            return
        if has_trajs:
            print(f'Trying {patcher.instance_id} again')
            os.remove(f'{run_path}/trajs/{patcher.instance_id}.json')
            if has_error:
                os.remove(f'{run_path}/errors/{patcher.instance_id}.error')
            if has_patches:
                os.remove(f'{run_path}/patches/{patcher.instance_id}_patches.json')
        print(f'Processing {patcher.instance_id} ...')
        patcher.query_LLM()
        patcher.evaluate_LLM()
        if patcher.requery_validator:
            patcher.evaluate_LLM()
        patcher.parse_LLM_solution()
        patcher.process_patches()
        
        if patcher.requery:
            print(f'requerying {patcher.instance_id}, attempt 1...')
            patcher.requery_LLM()
            patcher.parse_LLM_solution()
            patcher.process_patches()
        if patcher.requery:
            print(f'requerying {patcher.instance_id}, attempt 2...')
            patcher.requery_LLM()
            patcher.parse_LLM_solution()
            patcher.process_patches()
        if patcher.requery:
            print(f'requerying {patcher.instance_id}, attempt 3...')
            patcher.requery_LLM()
            patcher.parse_LLM_solution()
            patcher.process_patches()
        try:
            patcher.apply_patch()
            patcher.save_results(RUN_NAME)
        except Exception as e:
            print('Patch failed for instance_id', patcher.instance_id)
            print(e)
            with open(f"{run_path}/trajs/{patcher.instance_id}.json", "w") as f:
                f.write(json.dumps(patcher.message_queue, ensure_ascii=False, indent=4))
            with open(f"{run_path}/errors/{patcher.instance_id}.error", "w") as f:
                f.write(str(e))
            if len(patcher.patches) > 0:
                with open(f"{run_path}/patches/{patcher.instance_id}_patches.json", "w") as f:
                    f.write(json.dumps(patcher.patches[-1], ensure_ascii=False, indent=4))
        print(f'Finished {patcher.instance_id}.\n')
    except Exception as e:
        print('Something went wrong with patch', patcher.instance_id)
        print(e)
        with open(f"{run_path}/trajs/{patcher.instance_id}.json", "w") as f:
                        f.write(json.dumps(patcher.message_queue, ensure_ascii=False, indent=4))
        with open(f"{run_path}/errors/{patcher.instance_id}.error", "w") as f:
            f.write(str(e))
        if len(patcher.patches) > 0:
            with open(f"{run_path}/patches/{patcher.instance_id}_patches.json", "w") as f:
                f.write(json.dumps(patcher.patches[-1], ensure_ascii=False, indent=4))

with concurrent.futures.ThreadPoolExecutor(max_workers=2) as executor:
   executor.map(do_work, patchers) 

import json
## Create the jsonl necessary to run the benchmark
def save_patchers_info(name='CodeGPT'):
    path = f'../data/results/{name}/diffs'
    files =  [f for f in sorted(os.listdir(path)) if f.endswith('.diff')]
    jsonl_lines = []
    for f in files:
        instance_id = f.split('.diff')[0]
        try:
            with open(f"{path}/{f}", "r") as f:
                data = f.read()
            jsonl_line = {
                "instance_id": instance_id,
                "model_name_or_path": name,
                "model_patch": data
            }
            jsonl_lines.append(json.dumps(jsonl_line, ensure_ascii=False))
        
        except Exception as e:
            print(f'Failed to save info for {instance_id})')
            raise e
    with open(f"../data/results/{name}.jsonl", "w") as f:
        f.write("\n".join(jsonl_lines))
    print(f"Saved {len(files)} patchers to ../data/{name}.jsonl")
save_patchers_info(name=RUN_NAME)

predictions_path = os.path.abspath(f'../data/results/{RUN_NAME}.jsonl')
print(f'''python -m swebench.harness.run_evaluation \\
    --predictions_path {predictions_path} \\
    --max_workers 2 \\
    --run_id 0''')
